REFACTOR Removing the first test of divide that didn't check for result decimals.

+++ src/junittesting/MiniCalc.java
@@ -19,7 +19,7 @@ public class MiniCalc {
     }
 
     public double divide(int x1, int x2) {
-        return x1/x2;
+        return (double)x1/x2;
     }
     
 }
